hackwashu/requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.2
anyio==4.4.0
asgiref==3.7.2
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==24.2.0
Babel==2.15.0
bce-python-sdk==0.9.11
beautifulsoup4==4.11.1
bidict==0.22.1
Bootstrap-Flask==2.2.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
clib==0.0.1
click==8.1.3
click-plugins==1.1.1
cmake==3.25.0
contourpy==1.0.7
cssselect==1.2.0
cssutils==2.11.0
cycler==0.11.0
Cython==0.29.32
decorator==5.1.1
Django==5.0.2
dlib==19.23.0
dominate==2.7.0
easyocr==1.7.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.14.0
fire==0.6.0
Flask==2.2.2
flask-babel==4.0.0
Flask-Bootstrap==3.3.7.1
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Migrate==4.0.1
Flask-Minify==0.41
Flask-Reuploaded==1.3.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2024.5.0
future==1.0.0
gender-guesser==0.4.0
google-api-core==2.19.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-cloud-language==2.13.3
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gspread==5.9.0
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.4
image==1.5.33
imageio==2.34.1
imgaug==0.4.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsmin==3.0.1
kiwisolver==1.4.4
lazy_loader==0.4
lesscpy==0.15.1
lmdb==1.4.1
lxml==5.2.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.7.1
mediapipe==0.10.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.3
ninja==1.11.1.1
numpy==1.23.4
oauthlib==3.2.2
openai==0.27.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
osascript==2020.12.3
outcome==1.3.0.post0
packaging==23.1
paddleocr==2.7.3
paddlepaddle==2.6.1
pandas==1.5.3
pdf2docx==0.5.8
pillow==10.3.0
ply==3.11
premailer==3.10.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.20.0
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.0.9
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7
PyYAML==6.0.1
rapidfuzz==3.9.3
rarfile==4.2
rcssmin==1.1.1
regex==2024.5.15
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
runcmd==2020.12.3
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
selenium==4.23.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
sqlparse==0.4.4
sympy==1.12.1
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
tokenizers==0.19.1
torch==2.3.0
torchvision==0.18.0
tqdm==4.64.1
transformers==4.41.2
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==1.26.14
URLObject==2.4.3
visitor==0.1.3
visualdl==2.5.3
watchdog==4.0.1
websocket-client==1.8.0
Werkzeug==2.2.2
wsproto==1.2.0
WTForms==3.0.1
xlwt==1.3.0
xxhash==3.2.0
yarl==1.8.2
